Split identifier (UUID, name) related queries to own class.
